python/requirements.txt
openai==1.12.0
pandas==2.2.0
python-dotenv==1.0.0
